# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A minesweeper game API",
        "title": "Minesweeper API",
        "contact": {
            "name": "Christian Melgarejo",
            "email": "cmelgarejo.dev@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.8"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/v1/api/games": {
            "get": {
                "description": "Gets a list of games, only Admins can see it",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "game"
                ],
                "summary": "Gets a list of games",
                "parameters": [
                    {
                        "type": "string",
                        "default": "587fa65a9c375165828a6fbb5f9963a7",
                        "description": "API Key",
                        "name": "X-API-KEY",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.ResponseError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/responses.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.ResponseError"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a game of minesweeper and returns a gameID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "game"
                ],
                "summary": "Creates a game of minesweeper",
                "parameters": [
                    {
                        "type": "string",
                        "default": "587fa65a9c375165828a6fbb5f9963a7",
                        "description": "API Key",
                        "name": "X-API-KEY",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Game Input",
                        "name": "gameInput",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.GameCreateInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/engine.Game"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.ResponseError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/responses.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.ResponseError"
                        }
                    }
                }
            }
        },
        "/v1/api/games/start/{id}": {
            "post": {
                "description": "Starts a game of minesweeper and returns the mine field state",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "game"
                ],
                "summary": "Starts a game of minesweeper",
                "parameters": [
                    {
                        "type": "string",
                        "default": "ef99fdfd88565827ad330d83aac5fbaa",
                        "description": "Game ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "587fa65a9c375165828a6fbb5f9963a7",
                        "description": "API Key",
                        "name": "X-API-KEY",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.ResponseError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/responses.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.ResponseError"
                        }
                    }
                }
            }
        },
        "/v1/api/games/{id}": {
            "get": {
                "description": "Gets the information of a minesweeper game, fields and users",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "game"
                ],
                "summary": "Gets the information of a minesweeper game",
                "parameters": [
                    {
                        "type": "string",
                        "default": "ef99fdfd88565827ad330d83aac5fbaa",
                        "description": "Game ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "587fa65a9c375165828a6fbb5f9963a7",
                        "description": "API Key",
                        "name": "X-API-KEY",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/engine.Game"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.ResponseError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/responses.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.ResponseError"
                        }
                    }
                }
            },
            "patch": {
                "description": "Clicks field on a game of minesweeper and returns the mine field state",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "game"
                ],
                "summary": "Clicks field on a game of minesweeper",
                "parameters": [
                    {
                        "type": "string",
                        "default": "ef99fdfd88565827ad330d83aac5fbaa",
                        "description": "Game ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "587fa65a9c375165828a6fbb5f9963a7",
                        "description": "API Key",
                        "name": "X-API-KEY",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Game Input",
                        "name": "gameInput",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.GameInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.ResponseError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/responses.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.ResponseError"
                        }
                    }
                }
            }
        },
        "/v1/auth/signIn": {
            "post": {
                "description": "Sign in user of minesweeper and returns an API Key",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Sign in user of minesweeper",
                "parameters": [
                    {
                        "description": "Credentials",
                        "name": "credentials",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.Credentials"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.ResponseError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/responses.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.ResponseError"
                        }
                    }
                }
            }
        },
        "/v1/auth/user": {
            "post": {
                "description": "Creates an user of minesweeper and returns an userID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Creates an user of minesweeper",
                "parameters": [
                    {
                        "description": "User Input",
                        "name": "userInput",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.UserInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/responses.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.ResponseError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/responses.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.ResponseError"
                        }
                    }
                }
            }
        },
        "/v1/auth/user/{id}": {
            "get": {
                "description": "Gets the information of a minesweeper user, fields and users",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Gets the information of a minesweeper user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.ResponseError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/responses.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.ResponseError"
                        }
                    }
                }
            },
            "put": {
                "description": "Updates an user of minesweeper and returns an userID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Updates an user of minesweeper",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "User Input",
                        "name": "userInput",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.UserInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.ResponseError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/responses.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.ResponseError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "engine.Field": {
            "type": "object",
            "properties": {
                "adjMines": {
                    "description": "count of adjacent mines",
                    "type": "integer"
                },
                "clicked": {
                    "description": "indicated whether the field was clicked",
                    "type": "boolean"
                },
                "clickedBy": {
                    "description": "who clicked this field",
                    "type": "string"
                },
                "flagged": {
                    "description": "red flag in the field",
                    "type": "boolean"
                },
                "mine": {
                    "type": "boolean"
                },
                "position": {
                    "description": "position in the minefield",
                    "type": "object",
                    "$ref": "#/definitions/engine.Position"
                }
            }
        },
        "engine.Game": {
            "type": "object",
            "properties": {
                "cols": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "createdBy": {
                    "description": "who created this game",
                    "type": "string"
                },
                "finishedAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "mineField": {
                    "type": "array",
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/definitions/engine.Field"
                        }
                    }
                },
                "mines": {
                    "type": "integer"
                },
                "rows": {
                    "type": "integer"
                },
                "startedAt": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "engine.Position": {
            "type": "object",
            "properties": {
                "col": {
                    "description": "col of the field position",
                    "type": "integer"
                },
                "row": {
                    "description": "row of the field position",
                    "type": "integer"
                }
            }
        },
        "requests.Credentials": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string",
                    "example": "player1"
                },
                "username": {
                    "type": "string",
                    "example": "player1"
                }
            }
        },
        "requests.GameCreateInput": {
            "type": "object",
            "properties": {
                "col": {
                    "type": "integer",
                    "example": 5
                },
                "mines": {
                    "type": "integer",
                    "example": 5
                },
                "row": {
                    "type": "integer",
                    "example": 5
                }
            }
        },
        "requests.GameInput": {
            "type": "object",
            "properties": {
                "clickType": {
                    "type": "string",
                    "enum": [
                        "click",
                        "flag"
                    ]
                },
                "col": {
                    "type": "integer",
                    "example": 0
                },
                "row": {
                    "type": "integer",
                    "example": 0
                }
            }
        },
        "requests.UserInput": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "player1@player1.com"
                },
                "fullname": {
                    "type": "string",
                    "example": "Ready Player One"
                },
                "password": {
                    "type": "string",
                    "example": "player1"
                },
                "username": {
                    "type": "string",
                    "example": "player1"
                }
            }
        },
        "responses.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 12345
                },
                "details": {
                    "type": "string",
                    "example": "some more details"
                },
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/responses.ResponseError"
                    }
                },
                "message": {
                    "type": "string",
                    "example": "message"
                },
                "response": {
                    "type": "object"
                }
            }
        },
        "responses.ResponseError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 12345
                },
                "details": {
                    "type": "string",
                    "example": "some more details"
                },
                "message": {
                    "type": "string",
                    "example": "message"
                }
            }
        }
    }
}